CCCP-385, chore: move migration sequence enum to primitives
