+++ pallets/btc-registration-pool/src/lib.rs
@@ -4,7 +4,6 @@ pub mod migrations;
 mod pallet;
 pub mod weights;
 
-use frame_support::pallet_prelude::MaxEncodedLen;
 pub use pallet::pallet::*;
 use weights::WeightInfo;
 
@@ -64,29 +63,4 @@ pub struct VaultKeySubmission<AccountId> {
 	pub who: AccountId,
 	/// The generated public key. (33 bytes)
 	pub pub_key: Public,
-}
-
-/// Sequence of migrating registration pool.
-#[derive(
-	Eq,
-	PartialEq,
-	Ord,
-	PartialOrd,
-	Copy,
-	Clone,
-	Encode,
-	Decode,
-	RuntimeDebug,
-	TypeInfo,
-	MaxEncodedLen,
-	Default,
-)]
-pub enum MigrationSequence {
-	/// Normal sequence.
-	#[default]
-	Normal,
-	/// Prepare next system vault.
-	PrepareNextSystemVault,
-	/// Wait till all UTXOs transferred to the new system vault.
-	UTXOTransfer,
 }
